svc: Reserve space for handle instances ahead of time in WaitSynchronization
Rather than fill the vector with elements that we're immediately going
to overwrite, we can reserve the space ahead of time and push the
entries into the vector like normal.

This way we don't even need to zero out element before overwriting them.
While we're at it, we can also std::move the copies of the SharedPtr
instances that we're putting into the std::vector to avoid an
unnecessary atomic increment and decrement of the reference count.