service/lm: Implement SetDestination
This appears to be used as a means to redirect the log output on the
actual console. Because of that, the function is as simple as doing
nothing, since we always want the logging output to be visible. At most,
we can just log out the desired destination and leave it at that.